+++ gwcelery/celery.py
@@ -28,23 +28,21 @@ class Base:
 
     superevent_d_t_start = dict(gstlal=1.0,
                                 pycbc=1.0,
-                                mbtaonline=1.0,
-                                cwb=1.0,
-                                lib=1.0)
-    """Lower extent of superevent segments."""
+                                mbtaonline=1.0)
+    """Pipeline based lower extent of superevent segments.
+    For cwb and lib this is decided from extra attributes."""
+
+    superevent_d_t_end = dict(gstlal=1.0,
+                              pycbc=1.0,
+                              mbtaonline=1.0)
+    """Pipeline based upper extent of superevent segments
+    For cwb and lib this is decided from extra attributes."""
 
     superevent_query_d_t_start = 100.
     """Lower extent of superevents query"""
 
     superevent_query_d_t_end = 100.
     """Upper extent of superevents query"""
-
-    superevent_d_t_end = dict(gstlal=1.0,
-                              pycbc=1.0,
-                              mbtaonline=1.0,
-                              cwb=1.0,
-                              lib=1.0)
-    """Upper extent of superevent segments"""
 
     superevent_default_d_t_start = 1.0
     """Default lower extent of superevent segments"""