+++ gwcelery/tasks/superevents.py
@@ -35,18 +35,19 @@ def handle(payload):
     """
     gid = payload['uid']
     alert_type = payload['alert_type']
-    event_info = _event_info(payload)
 
     try:
         far = payload['object']['far']
     except KeyError:
         log.info(
-            'skipping %s because LVAlert message does not provide FAR', gid)
+            'Skipping %s because LVAlert message does not provide FAR', gid)
         return
     else:
         if far > app.conf['superevent_far_threshold']:
-            log.info("Skipping processing of %s because of low far", gid)
+            log.info("Skipping processing of %s because of low FAR", gid)
             return
+
+    event_info = _get_event_info(payload)
 
     query_times = (event_info['gpstime'] -
                    app.conf['superevent_query_d_t_start'],
@@ -55,7 +56,6 @@ def handle(payload):
 
     sid, preferred_flag, superevents = gracedb.get_superevent(
         gid, query='''{} .. {}'''.format(*query_times))
-    superevents = _superevent_segment_list(superevents)
 
     d_t_start, d_t_end = _get_dts(event_info)
 
@@ -63,36 +63,47 @@ def handle(payload):
     if sid is None and alert_type == 'new':
         log.debug('Entered 1st if')
         event_segment = _Event(event_info['gpstime'],
-                               event_info['uid'],
+                               event_info['gpstime'] - d_t_start,
+                               event_info['gpstime'] + d_t_end,
+                               event_info['graceid'],
                                event_info['group'],
                                event_info['pipeline'],
                                event_info['search'],
                                event_dict=payload)
 
-        # Check which superevent window trigger gpstime falls in
-        if superevents.intersects_segment(event_segment):
-            # FIXME .find() has to be generalized when event_segments
-            # have a chance of not completely being contained in a
-            # superevent segment
-            superevent = superevents[superevents.find(event_segment)]
-            log.info('Event %s in window of %s',
-                     gid, superevent.superevent_id)
-            gracedb.add_event_to_superevent(superevent.superevent_id,
-                                            event_segment.gid)
-            # ADDITIONAL LOGIC HERE IF THE TIME
-            # WINDOW IS TO BE CHANGED BASED ON
-            # TRIGGER PARAMETERS
-            _update_preferred_event(superevent.superevent_id,
-                                    superevent.preferred_event,
-                                    event_segment.gid)
-        # Create a new event if not in any time window
-        else:
-            gracedb.create_superevent(event_info['uid'],
+        superevent = _partially_intersects(superevents, event_segment)
+
+        if not superevent:
+            log.info("New event %s with no superevent in GraceDb, \
+                     creating new superevent", gid)
+            gracedb.create_superevent(event_info['graceid'],
                                       event_info['gpstime'],
                                       d_t_start,
                                       d_t_end)
+            return
 
-    # Condition 2/2
+        log.info('Event %s in window of %s. Adding event to superevent',
+                 gid, superevent.superevent_id)
+        gracedb.add_event_to_superevent(superevent.superevent_id,
+                                        event_segment.gid)
+        # extend the time window of the superevent
+        new_superevent = superevent | event_segment
+        if new_superevent is not superevent:
+            log.info("%s not completely contained in %s, "
+                     "extending superevent window",
+                     event_segment.gid, superevent.superevent_id)
+            new_t_start, new_t_end = new_superevent[0], new_superevent[1]
+        else:
+            log.info("%s is completely contained in %s",
+                     event_segment.gid, superevent.superevent_id)
+            new_t_start = new_t_end = None
+
+        _update_superevent(superevent.superevent_id,
+                           superevent.preferred_event,
+                           event_info,
+                           t_start=new_t_start,
+                           t_end=new_t_end)
+
     elif sid and alert_type == 'new':
         # ERROR SITUATION
         log.debug('3rd if: SID returned and new alert')
@@ -103,49 +114,63 @@ def handle(payload):
         log.critical('Unhandled by parse_trigger, passing...')
 
 
-def _event_info(payload):
-    """Helper function to fetch required event info from GraceDb
+def _get_event_info(payload):
+    """Helper function to fetch required event info (from GraceDb)
     at once and reduce polling
     """
-    alert_type = payload['alert_type']
-    if alert_type == 'new':
-        event_info = dict(uid=payload['uid'],
-                          gpstime=payload['object']['gpstime'],
-                          far=payload['object']['far'],
-                          group=payload['object']['group'],
-                          pipeline=payload['object']['pipeline'],
-                          search=payload['object'].get('search'),
-                          alert_type=payload['alert_type'])
-    else:
-        event_dict = gracedb.get_event(payload['uid'])
-        event_info = dict(uid=event_dict['graceid'],
-                          gpstime=event_dict['gpstime'],
-                          far=event_dict['far'],
-                          group=event_dict['group'],
-                          pipeline=event_dict['pipeline'],
-                          search=event_dict.get('search'),
-                          alert_type=payload['alert_type'])
+    # pull basic info
+    event_info = dict(
+        graceid=payload['object']['graceid'],
+        gpstime=payload['object']['gpstime'],
+        far=payload['object']['far'],
+        group=payload['object']['group'],
+        pipeline=payload['object']['pipeline'],
+        search=payload['object'].get('search'),
+        alert_type=payload['alert_type'])
+    # pull pipeline based extra attributes
+    if payload['object']['pipeline'].lower() == 'cwb':
+        extra_attributes = ['duration', 'start_time']
+        event_info.update(
+            {attr:
+             payload['object']['extra_attributes']['MultiBurst'][attr]
+             for attr in extra_attributes})
+    elif payload['object']['pipeline'].lower() == 'lib':
+        extra_attributes = ['quality_mean', 'frequency_mean']
+        event_info.update(
+            {attr:
+             payload['object']['extra_attributes']['LalInferenceBurst'][attr]
+             for attr in extra_attributes})
+    # if required, add CBC specific extra_attributes here
     return event_info
 
 
 def _get_dts(event_info):
     """
-    Returns the dt_start and dt_end values based on CBC/Burst
-    for new and update type alerts
+    Returns the d_t_start and d_t_end values based on CBC/Burst
+    type alerts
     """
     group = event_info['group']
-    dt_start = app.conf['superevent_d_t_start'].get(
-        group.lower(), app.conf['superevent_default_d_t_start'])
+    pipeline = event_info['pipeline']
+    if pipeline.lower() == 'cwb':
+        d_t_start = d_t_end = event_info['duration']
+    elif pipeline.lower() == 'lib':
+        d_t_start = d_t_end \
+            = event_info['quality_mean']/event_info['frequency_mean']
+    elif group.lower() == 'cbc':
+        d_t_start = app.conf['superevent_d_t_start'][pipeline.lower()]
+        d_t_end = app.conf['superevent_d_t_end'][pipeline.lower()]
+    else:
+        d_t_start = app.conf['superevent_default_d_t_start']
+        d_t_end = app.conf['superevent_default_d_t_end']
+    return d_t_start, d_t_end
 
-    dt_end = app.conf['superevent_d_t_end'].get(
-        group.lower(), app.conf['superevent_default_d_t_end'])
 
-    return dt_start, dt_end
-
-
-def _update_preferred_event(superevent_id, preferred_event, gid):
+def _update_superevent(superevent_id, preferred_event, new_event_dict,
+                       t_start, t_end):
     """
-    Update superevent with the new trigger id based on FAR values.
+    Update preferred event and/or change time window.
+    Decision between `preferred_event` and `new_event`
+    based on FAR values.
 
     Parameters
     ----------
@@ -153,18 +178,32 @@ def _update_preferred_event(superevent_id, preferred_event, gid):
         superevent uid
     preferred_event : str
         preferred event id of the superevent
-    gid : str
-        uid of the new trigger
+    new_event_dict : dict
+        event info of the new trigger as a dictionary
+    t_start : float
+        start time of `superevent_id`, None for no change
+    t_end : float
+        end time of `superevent_id`, None for no change
     """
-    new_event = gracedb.client.event(gid).json()
-    preferred_event = gracedb.client.event(preferred_event).json()
-    if new_event['far'] < preferred_event['far']:
-        gracedb.update_superevent(superevent_id, preferred_event=gid)
+    preferred_event_dict = gracedb.client.event(preferred_event).json()
+
+    if new_event_dict['far'] < preferred_event_dict['far']:
+        gracedb.update_superevent(superevent_id,
+                                  preferred_event=new_event_dict['graceid'],
+                                  t_start=t_start,
+                                  t_end=t_end)
+    else:
+        # no action needed if t_start or t_end is not supplied
+        if t_start or t_end:
+            gracedb.update_superevent(superevent_id,
+                                      t_start=t_start,
+                                      t_end=t_end)
 
 
 def _superevent_segment_list(superevents):
     """Ingests a list of superevent dictionaries, and returns
-    a segmentlist of the same
+    a segmentlist with start and end times as the duration of
+    each segment
 
     Parameters
     ----------
@@ -177,21 +216,51 @@ def _superevent_segment_list(superevents):
     superevent_list : segmentlist
         superevents as a segmentlist object
     """
-    superevent_list = \
-        [_SuperEvent(s['t_0'], s['t_start'], s['t_end'],
-                     s['superevent_id'], s['preferred_event'], s)
-            for s in superevents]
-    return segmentlist(superevent_list)
+    return segmentlist([_SuperEvent(s.get('t_start'),
+                        s.get('t_end'),
+                        s.get('t_0'),
+                        s.get('superevent_id'),
+                        s.get('preferred_event'),
+                        s)
+                       for s in superevents])
+
+
+def _partially_intersects(superevents, event_segment):
+    """Similar to :meth:`segmentlist.find`
+    except it also returns the segment of
+    `superevents` which partially intersects argument.
+    If there are more than one intersections,
+    first occurence is returned.
+
+    Parameters
+    ----------
+    superevents : list
+        list pulled down using the gracedb client
+        :method:`superevents`
+    event_segment : segment
+        segment object whose index is wanted
+
+    Returns
+    -------
+    match_segment   : segment
+        segment in `self` which intersects. `None` if not found
+    """
+    # create a segmentlist using start and end times
+    superevents = _superevent_segment_list(superevents)
+    for s in superevents:
+        if s.intersects(event_segment):
+            return s
+    return None
 
 
 class _Event(segment):
     """An event implemented as an extension of
-    :class:`glue.segments.segment`
+    :class:`segment`
     """
-    def __new__(cls, t0, *args, **kwargs):
-        return super().__new__(cls, t0, t0)
+    def __new__(cls, t0, t_start, t_end, *args, **kwargs):
+        return super().__new__(cls, t_start, t_end)
 
-    def __init__(self, t0, gid, group=None, pipeline=None,
+    def __init__(self, t0, t_start, t_end, gid, group=None, pipeline=None,
                  search=None, event_dict={}):
         self.t0 = t0
         self.gid = gid
@@ -203,7 +272,7 @@ class _Event(segment):
 
 class _SuperEvent(segment):
     """An superevent implemented as an extension of
-    :class:`glue.segments.segment`
+    :class:`segment`
     """
     def __new__(cls, t_start, t_end, *args, **kwargs):
         return super().__new__(cls, t_start, t_end)